backend/requirements.txt
flask
flask-cors
openai-whisper
groq
sarvamai
python-dotenv
firebase-admin
